Add files via upload
"BounceBall" is a skipping ball game created using Android Studio. In this game, players control the bouncing ball with the objective of keeping the ball on the ground. In the game, players control a bouncing ball, navigating it through various obstacles and challenges. The objective is to keep the ball bouncing for as long as possible while avoiding obstacles such as pits, spikes, or moving objects. Players typically use touch controls to guide the ball's movement, tapping or swiping on the screen to direct it left or right. As players progress.